+++ src/main/java/InterestService/InterestsMechanism.java
@@ -1,5 +1,5 @@
 package InterestService;
 
 public interface InterestsMechanism {
-    double generateInterest(int amountOfMoney, double percentage, short howManyCapitalisation);
+    double generateInterest(int amountOfMoney, double percentage, int howManyCapitalisation);
 }
