+++ src/main/java/InterestService/InvestmentInterestMechanism.java
@@ -2,7 +2,7 @@ package InterestService;
 
 public class InvestmentInterestMechanism implements InterestsMechanism {
 
-    public double generateInterest(int amountOfMoney, double percentage, short howManyCapitalisation) {
+    public double generateInterest(int amountOfMoney, double percentage, int howManyCapitalisation) {
         return amountOfMoney * howManyCapitalisation * percentage;
     }
 }
